+++ packages/shared/constants.js
@@ -1,0 +1,21 @@
+/**
+ * Copyright (c) Areslabs.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+
+export const originElementAttrName = '_o'
+export const viewOrigin = "1"
+export const imageOrigin = "2"
+export const innerTextOrigin = "3"
+export const outerTextOrigin = "4"
+export const errorViewOrigin = "5"
+export const touchableWithoutFeedbackOrigin = "6"
+export const touchableOpacityOrigin = "7"
+export const touchableHighlightOrigin = "8"
+
+
+export const reactFragmentFlag = "_FF"